Extracting payload serde code from AbstractBinder into a utility class
Fix #1030
